requirements.txt
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
google-api-core==1.17.0
google-api-python-client==1.8.4
google-api-support==0.0.4
google-auth==1.16.0
google-auth-httplib2==0.0.3
google-cloud-texttospeech==2.0.0
googleapis-common-protos==1.51.0
grpcio==1.29.0
httplib2==0.18.1
idna==2.9
imageio==2.8.0
imageio-ffmpeg==0.4.2
libcst==0.3.6
moviepy==1.0.3
mypy-extensions==0.4.3
numpy==1.18.4
oauth2client==4.1.3
oauthlib==3.1.0
pandas==1.0.4
pdf2image==1.13.1
Pillow==7.1.2
proglog==0.1.9
proto-plus==0.4.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
selenium==3.141.0
six==1.15.0
tqdm==4.46.1
tweepy==3.8.0
tweet-preprocessor==0.6.0
typing-extensions==3.7.4.2
typing-inspect==0.6.0
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.9
